==> Part1_Data_cleaning.ipynb <==
# Part I. Data cleaning

#import packages
import pandas as pd
import numpy as np

## User data

user_data = pd.read_csv('data_format1/user_log_format1.csv')
user_data.head()

user_data.shape

#find out the missing value in this file
user_data.isnull().sum()

#get the mode brand id for all stores, fillin NA value in brain_id column
missing = user_data[user_data.brand_id.isnull()].index
seller = user_data.groupby(['seller_id']).apply(lambda x:x.brand_id.mode()[0]).reset_index()
get_brand = user_data.loc[missing]
get_brand = get_brand.merge(seller,how='left',on=['seller_id'])[0].astype('float32')
get_brand.index = missing
user_data.loc[missing,'brand_id'] = get_brand

#see whether null values are filled
user_data.isnull().sum()

#unique customer purchase history
user_data['user_id'].nunique()

user_data.shape

user_data.to_csv('cleaned_raw/user_data.csv')

## User info

#user info
user_info = pd.read_csv('data_format1/user_info_format1.csv')
user_info.head()

#user info
user_info.isnull().sum()

user_info.age_range.fillna(user_info.age_range.median(),inplace=True)
user_info.gender.fillna(2,inplace=True)
#user info NA value exam
user_info.isnull().sum()

user_info = user_info[user_info['age_range'] != 0]

df_age = pd.get_dummies(user_info.age_range,prefix='age')
df_gender = pd.get_dummies(user_info.gender)
df_gender.rename(columns={0:'female',1:'male',2:'unknown'},inplace=True)
user_info = pd.concat([user_info.user_id, df_age, df_gender], axis=1)
del df_age, df_gender

user_info.to_csv('cleaned_raw/user_info.csv')
