==> Part2.2_product_diversity.ipynb <==
# Part 2 Feature Engineering

## 2.2 Product diversity

import pandas as pd
import numpy as np

user_data = pd.read_csv('cleaned_raw/user_data.csv')

user_data.head()

### 2.2.1 merchant
for each merchant, obtain the count of the 4 different actions and sum of them for unique item/brand/category

#merchant action dataframe prep
mc_action = user_data[['seller_id','brand_id','cat_id','item_id','action_type']]
action_dummies = pd.get_dummies(mc_action['action_type'])
mc_action = pd.concat([mc_action, action_dummies], axis=1)
mc_action.rename(columns={0: "click", 1:"add", 2:"buy", 3:"save"},inplace=True)
del action_dummies
mc_action.head()

# for each unique item, get the overall count
mc_item_overall = mc_action.groupby(['seller_id','item_id'], as_index=False)['click','add','buy','save'].sum()

mc_item_overall['total_action'] = mc_item_overall['click'] + mc_item_overall['add'] + mc_item_overall['buy'] + mc_item_overall['save']
mc_item_overall.head()

# for each unique cat, get the overall count
mc_cat_overall = mc_action.groupby(['seller_id','cat_id'], as_index=False)['click','add','buy','save'].sum()

mc_cat_overall['total_action'] = mc_cat_overall['click'] + mc_cat_overall['add'] + mc_cat_overall['buy'] + mc_cat_overall['save']
mc_cat_overall.head()

# for each unique brand, get the overall count
mc_brand_overall = mc_action.groupby(['seller_id','brand_id'], as_index=False)['click','add','buy','save'].sum()

mc_brand_overall['total_action'] = mc_brand_overall['click'] + mc_brand_overall['buy'] + mc_brand_overall['save']
mc_brand_overall.head()

### 2.2.2 user
for each merchant, obtain the count of the 4 different actions and sum of them for unique item/brand/category

#user action dataframe prep
user_action = user_data[['seller_id','user_id','brand_id','cat_id','item_id','action_type']]
action_dummies = pd.get_dummies(user_action['action_type'])
user_action = pd.concat([user_action, action_dummies], axis=1)
user_action.rename(columns={0: "click", 1:"add", 2:"buy", 3:"save"},inplace=True)
del action_dummies
user_action.head()

# for each unique item, get the overall count
user_item_overall = user_action.groupby(['user_id','item_id'], as_index=False)['click','add','buy','save'].sum()
user_item_overall['total_action'] = user_item_overall['click'] + user_item_overall['add'] + user_item_overall['buy'] + user_item_overall['save']
user_item_overall.head()

# for each unique cat, get the overall count
user_cat_overall = user_action.groupby(['user_id','cat_id'], as_index=False)['click','add','buy','save'].sum()
user_cat_overall['total_action'] = user_cat_overall['click'] + user_cat_overall['add'] + user_cat_overall['buy'] + user_cat_overall['save']
user_cat_overall.head()

# for each unique brand, get the overall count
user_brand_overall = user_action.groupby(['user_id','brand_id'], as_index=False)['click','add','buy','save'].sum()
user_brand_overall['total_action'] = user_cat_overall['click'] + user_cat_overall['add'] + user_cat_overall['buy'] + user_cat_overall['save']
user_brand_overall.head()

### 2.2.3 <merchant, user>
for each unique <merchant,user>, obtain the count of the 4 different actions and sum of them for unique item/brand/category

# for each unique item, get the overall count
mu_item_overall = user_action.groupby(['seller_id','user_id','item_id'], as_index=False)['click','add','buy','save'].sum()
mu_item_overall['total_action'] = mu_item_overall['click'] + mu_item_overall['add'] + mu_item_overall['buy'] + mu_item_overall['save']
mu_item_overall.head()

# for each unique cat, get the overall count
mu_cat_overall = user_action.groupby(['seller_id','user_id','cat_id'], as_index=False)['click','add','buy','save'].sum()
mu_cat_overall['total_action'] = mu_cat_overall['click'] + mu_cat_overall['add'] + mu_cat_overall['buy'] + mu_cat_overall['save']
mu_cat_overall.head()

# for each unique brand, get the overall count
mu_brand_overall = user_action.groupby(['seller_id','user_id','brand_id'], as_index=False)['click','add','buy','save'].sum()
mu_brand_overall['total_action'] = mu_brand_overall['click'] + mu_brand_overall['add'] + mu_brand_overall['buy'] + mu_brand_overall['save']
mu_brand_overall.head()
