==> Part2.4_market_share.ipynb <==
# Part 2. Feature Engineering

## 2.3 Market share feature

### 2.3.1 get unique count for item/cat/brand for each seller

#import packages
import numpy as np
import pandas as pd

user_data = pd.read_csv('cleaned_raw/user_data.csv')

def Categorize(n):
    if n == 1111:
        return 2
    elif n >= 1104:
        return 1
    else:
        return 0
user_data['sort_time'] = user_data.time_stamp.map(Categorize)

def unique_count_seller(id_name):
    col_id = id_name + '_id'
    col_name = id_name + '_count'
    
    count = user_data[['seller_id',col_id]]
    count = count.groupby(['seller_id'])[col_id].nunique().reset_index()
    count.columns = ['seller_id', col_name]
    return (count)

item_count = unique_count_seller('item')
cat_count = unique_count_seller('cat')
brand_count = unique_count_seller('brand')

### 2.3.2 get repeat purchase user for every seller before double 11 (Nov. 11th); we could also get a label for users: whether or not they are repeat buyers (regardless of the seller)

repeat_purchase = user_data[(user_data['action_type'] == 2) & (user_data['time_stamp'] < 1111)]
repeat_purchase = repeat_purchase.groupby(['seller_id'])['user_id'].value_counts().to_frame()
repeat_purchase.columns = ['purchase_count']
repeat_purchase.reset_index(inplace=True)
repeat_purchase.head()

repeat_purchase = repeat_purchase[repeat_purchase['purchase_count'] > 1]
repeat_purchase = repeat_purchase.groupby(['seller_id'])['user_id'].nunique().reset_index()

repeat_purchase.describe()

# Because some sellers without repeated purchases are excluded,we have to add them and fill 0 for future merge.
all_sellerid = pd.DataFrame({'seller_id':range(1, 4996 ,1)})
repeat_purchase = pd.merge(all_sellerid,repeat_purchase,how='left',on=['seller_id']).fillna(0)
repeat_purchase.head()

repeat_purchase.columns = ['seller_id', 'repeat_users_count']
repeat_purchase.head()

### 2.3.2 seller analysis
I: count each action click/add/buy/save/total for each seller and proportion/mean  
II: count each action done by each user in each seller

# seller_id and action_type
actions = user_data[['seller_id', 'action_type']]
actions.head()

# extract each action type
dummy_action = pd.get_dummies(actions['action_type'], prefix='action')
dummy_action.head()

# count the number of each actions for each seller
seller_actions = pd.concat([actions,dummy_action], axis=1).groupby(['seller_id'], as_index=False).sum()
#.drop('action_type', axis=1, inplace=True)

seller_actions.drop('action_type',1, inplace=True)
seller_actions['sellerTotalAction'] = seller_actions.apply(lambda x: x["action_0"] + x["action_1"] + x["action_2"] + x["action_3"], axis=1)
seller_actions.head()

#Clicked to buy conversion rate, added to shopping cart purchase conversion rate, purchased conversion rate
seller_actions['click_to_buy_ratio'] = np.log1p(seller_actions['action_2']) - np.log1p(seller_actions['action_0'])
# add-in-cart to purchase conversion rate, purchased conversion rate
seller_actions['cart_to_buy_ratio'] = np.log1p(seller_actions['action_2']) - np.log1p(seller_actions['action_1'])
# add-in-fav to purchase conversion rate
seller_actions['fav_to_buy_ratio'] = np.log1p(seller_actions['action_2']) - np.log1p(seller_actions['action_3'])

seller_actions.head()

# user_id, seller_id and action_type
user_actions = user_data[['seller_id', 'user_id', 'action_type']]
dummy_user = pd.get_dummies(user_actions['action_type'], prefix='seller_user_action')
dummy_user.head()

su_action = pd.concat([user_actions, dummy_user], axis=1)
su_action.drop('action_type',1, inplace=True)
su_action.head()

# drop duplicates so that we can count the number of users who made actions for each seller
su_action.drop_duplicates(inplace=True)
dummy_user_1 = su_action.groupby(['seller_id'])['seller_user_action_0'].sum()
dummy_user_2 = su_action.groupby(['seller_id'])['seller_user_action_1'].sum()
dummy_user_3 = su_action.groupby(['seller_id'])['seller_user_action_2'].sum()
dummy_user_4 = su_action.groupby(['seller_id'])['seller_user_action_3'].sum()

su_action = pd.concat([dummy_user_1, dummy_user_2, dummy_user_3, dummy_user_4], axis=1).reset_index()
su_action.head()

seller_actions = pd.merge(seller_actions,item_count,on=['seller_id'])
seller_actions = pd.merge(seller_actions,cat_count,on=['seller_id'])
seller_actions = pd.merge(seller_actions,brand_count,on=['seller_id'])
seller_actions = pd.merge(seller_actions,repeat_purchase,on=['seller_id'])
seller_actions = pd.merge(seller_actions,su_action,on=['seller_id'])

seller_actions.head()

# calculate the ratio of item, category, brand of each seller of total to see scale of each seller
seller_actions.item_ratio = seller_actions.item_count/user_data.item_id.nunique()
seller_actions.cat_ratio = seller_actions.cat_count/user_data.cat_id.nunique()
seller_actions.brand_ratio = seller_actions.brand_count/user_data.brand_id.nunique()

# calculate the proportion of clicks, add-to-cart, purchase, add-to-fav of each seller of total of each activities 
# in the whole market
seller_actions.su_click_ratio = seller_actions.seller_user_action_0/user_data[user_data.action_type == 0]['user_id'].nunique()
seller_actions.su_cart_ratio = seller_actions.seller_user_action_1/user_data[user_data.action_type == 1]['user_id'].nunique()
seller_actions.su_buy_ratio = seller_actions.seller_user_action_2/user_data[user_data.action_type == 2]['user_id'].nunique()
seller_actions.su_fav_ratio = seller_actions.seller_user_action_3/user_data[user_data.action_type == 3]['user_id'].nunique()

seller_actions.head()

seller_actions.rename(columns={'seller_user_action_0': "su_click", 'seller_user_action_1':"su_cart",
                                   "seller_user_action_2":"su_buy", "seller_user_action_3":"su_fav", 
                                   "action_0":"click", "action_1":"cart", "action_2":"buy", "action_3":"fav"},inplace=True)

seller_actions.head()

seller_double11 = pd.read_csv('data_output/Double11_seller.csv')

seller_actions = pd.merge(seller_actions,seller_double11,on=['seller_id'])
seller_actions.shape

seller_actions.to_csv('data_output/seller_features.csv', index=False)
