==> Part2.5_repeat_behavior_related.ipynb <==
# Part 2 Feature Engineering

## 2.5 Repeat behavior

import pandas as pd
import numpy as np

user_data = pd.read_csv('cleaned_raw/user_data.csv')
user_data.head()

### 2.5.1 For merchants: how many repeat buyers? What is the ratio of repeat buyer vs unique buyer for a certain merchant?  These are the features reflect the level of attractiveness 

repeat_purchase = user_data[(user_data['action_type'] == 2) & (user_data['time_stamp'] < 1111)]
repeat_purchase = repeat_purchase.groupby(['seller_id'])['user_id'].value_counts().to_frame()
repeat_purchase.columns = ['purchase_count']
repeat_purchase.reset_index(inplace=True)
repeat_purchase.head()

repeat_purchase_True = repeat_purchase[repeat_purchase['purchase_count'] > 1]
repeat_buyer_label = repeat_purchase_True[['user_id']].drop_duplicates()
repeat_count_each_seller = repeat_purchase_True.groupby(['seller_id'])['user_id'].nunique().reset_index()
repeat_count_each_seller.columns = ['seller_id', 'repeat_user_num']
all_seller_id = pd.DataFrame({'seller_id':range(1, 4997 ,1)})
repeat_count_each_seller = pd.merge(all_seller_id, repeat_count_each_seller, how='left',on=['seller_id']).fillna(0)

**this could reflect the attraction to repeat buyers**

repeat_count_each_seller.head()

all_unique_user = repeat_purchase.groupby(['seller_id'])['user_id'].count().reset_index()
repeat_m = pd.merge(repeat_count_each_seller, all_unique_user, on='seller_id' )
repeat_m = repeat_m[['seller_id','repeat_user_num','user_id']]
repeat_m = repeat_m.rename(columns={'user_id':'all_unique_user'})

repeat_m.dtypes

repeat_m['att_ratio'] = repeat_m['repeat_user_num'] / repeat_m['all_unique_user'].astype(float)

repeat_m.shape

repeat_m.to_csv('data_processed/repeat_merchant_result.csv')

### 2.5.1 For users: in how many stores did they do repeat buying? What is the average repeat buying action interval?

repeat_user = user_data[(user_data['action_type'] == 2) & (user_data['time_stamp'] < 1111)]

repeat_user.head()

repeat_user = repeat_user.sort_values(['user_id','seller_id','time_stamp']).reset_index(drop=True)
repeat_user.head()

repeat_buyer_ls = repeat_purchase_True['user_id'].unique()

repeat_buyer_ls

repeat_user_t = repeat_user.loc[repeat_user['user_id'].isin(repeat_buyer_ls)]

repeat_user_t.shape

repeat_user.shape

repeat_user_t.head()

#repeat_user_se = repeat_user_t.head(100)

repeat_user_t['diff'] = repeat_user_t.groupby(['user_id','seller_id'])['time_stamp'].diff()

repeat_user_t1 = repeat_user_t.dropna(subset=['diff'])
repeat_user_t1.shape

repeat_user_t1.head()

conditions =[
    (repeat_user_t1['diff'] < 35.0),
    (repeat_user_t1['diff'] >= 35.0) & (repeat_user_t1['diff'] < 135.0),
    (repeat_user_t1['diff'] >= 135.0) & (repeat_user_t1['diff'] < 235.0),
    (repeat_user_t1['diff'] >= 235.0) & (repeat_user_t1['diff'] < 335.0),
    (repeat_user_t1['diff'] >= 335.0) & (repeat_user_t1['diff'] < 435.0),
    (repeat_user_t1['diff'] >= 435.0) & (repeat_user_t1['diff'] < 535.0),
    (repeat_user_t1['diff'] >= 535.0) & (repeat_user_t1['diff'] < 635.0),
]

choices = [repeat_user_t1['diff'], repeat_user_t1['diff']-70,  repeat_user_t1['diff']-140,
           repeat_user_t1['diff']-210, repeat_user_t1['diff']-280, repeat_user_t1['diff']-350,
          repeat_user_t1['diff']-420]

repeat_user_t1['diff_new'] = np.select(conditions, choices, default=0)

repeat_user_result = repeat_user_t1[['user_id','seller_id', 'diff_new']].reset_index(drop=True)

repeat_user_result.head()

repeat_user_result = pd.DataFrame(repeat_user_result.groupby(['user_id', 'seller_id'])['diff_new'].mean().reset_index())

repeat_user_result.to_csv('data_processed/repeat_user_result.csv')
