==> Part3.1_Data Merge.ipynb <==
# Part 3. Data Merge

## Outline:
### 3.1 User merge
- gender
- age
- user action analysis
- double 11 user analysis
- repeat user:
    ` repeat behavior interval (regardless of seller)
    ` repeat behavior count (regardless of seller)
    ` repeat user attribute (binary variable 0-1)  
    
### 3.2 Seller merge
- seller action analysis
- double 11 seller analysis
- repeat seller (how many repeat buyers & attractive ratio for each seller)  

### 3.3 <seller, user> merge
for each unique pair of seller and user
- <seller, user> action analysis
- double 11 <seller, user>  analysis
- <seller, user> product activity analysis
- repeat <seller, user> (how many times of repeat purchase happend?)

### 3.4 merge all data with test, and split the train test with s=0.67

import pandas as pd

## 3.2 Seller merge

seller_features = pd.read_csv("data_processed/seller_features.csv")
seller_features.head()

seller_overall = pd.read_csv('data_processed/seller_overall.csv')
seller_overall.head()

seller_repeat = pd.read_csv('data_processed/repeat_merchant_result.csv')
seller_repeat.head()

seller_features.shape

seller_overall.shape

seller_repeat.shape

seller_FINAL = pd.merge(seller_features, seller_overall, on='seller_id')
seller_FINAL = pd.merge(seller_FINAL, seller_repeat, how='left',on='seller_id').fillna(0)

seller_FINAL.shape

### column cleaning

seller_FINAL.columns

seller_FINAL = seller_FINAL.drop(['Unnamed: 0_x','Unnamed: 0_y'], axis=1)

#data output
seller_FINAL.to_csv('Final_data/seller_FINAL.csv')

## 3.3 <seller, user> merge

su_overall = pd.read_csv("data_processed/su_overall.csv")
su_overall.head()

su_double11 = pd.read_csv("data_processed/Double11_su.csv")
su_double11.head()

su_pd = pd.read_csv("data_processed/su_pd.csv")
su_pd.head()

su_repeat = pd.read_csv("data_processed/su_repeat.csv")
su_repeat.head()

su_FINAL = pd.merge(su_overall, su_double11, how ='left', on='su_id')

su_FINAL = pd.merge(su_FINAL, su_pd, how ='left', on='su_id')

su_FINAL = pd.merge(su_FINAL, su_repeat, how ='left', on='su_id')

### column cleaning

su_FINAL.columns

su_FINAL = su_FINAL.drop(['Unnamed: 0_x','Unnamed: 0_y', 'Unnamed: 0'], axis=1)

#data output
su_FINAL.to_csv('Final_data/su_FINAL.csv')
