==> Part3.2_Data Output.ipynb <==
### 3.4 merge all data with test

import pandas as pd

label = pd.read_csv('data_format1/train_format1.csv')
label.head()


label.rename(columns = {"merchant_id":"seller_id"}, inplace=True)
label['user_id'] = label['user_id'].astype(str)
label['seller_id'] = label['seller_id'].astype(str)
label['su_id'] = label['user_id'].str.cat(label['seller_id'], sep='_')
label.head()

label.to_csv('Final_data/label_edited.csv')

su = pd.read_csv('Final_data/su_FINAL.csv')
seller = pd.read_csv('Final_data/seller_FINAL.csv')
user = pd.read_csv('Final_data/user_FINAL.csv')

from sklearn.model_selection import train_test_split
y = label['label']
X = label
X_train, X_test, y_train, y_test = train_test_split(X, y,
                                                    stratify=y, 
                                                    test_size=0.5)

X_train.shape

X_train = pd.merge(X_train, su, how='left', on='su_id')

seller['seller_id'] = seller['seller_id'].astype(str)
X_train['seller_id'] = X_train['seller_id'].astype(str)
user['user_id'] = user['user_id'].astype(str)
X_test['seller_id'] = X_test['seller_id'].astype(str)

X_train = pd.merge(X_train, seller, how='left', on='seller_id')

X_train = pd.merge(X_train, user, how='left', on='user_id')

X_test = pd.merge(X_test, su, how='left', on='su_id')
X_test = pd.merge(X_test, seller, how='left', on='seller_id')
X_test = pd.merge(X_test, user, how='left', on='user_id')

X_train.to_csv('Final_data/X_train.csv')
X_test.to_csv('Final_data/X_test.csv')
y_train.to_csv('Final_data/y_train.csv')
y_test.to_csv('Final_data/y_test.csv')
